refactor(optimize): extract walk-forward data models
walk_forward.py (861 lines) mixed the WFA data contracts (WalkForwardConfig,
WalkForwardWindow, WindowResult, WalkForwardResult) with the
WalkForwardOptimizer. Move the contracts into walk_forward_models.py (861->765)
and re-export. Behavior-preserving; 145 optimize tests pass (optuna installed
locally to verify).
